requirements.txt
# Core LangChain and AI
langchain-google-genai>=1.0.0
langchain>=0.1.0
langgraph>=0.4.0

# Graph Database
neo4j>=5.0.0

# Data Models
pydantic>=2.0.0

# Environment Management
python-dotenv>=1.0.0

# HTML Processing
beautifulsoup4>=4.12.0

# Optional: For development
jupyter>=1.0.0 